Add files via upload
The purpose of the validator and aim of the "validator culture" is to guide logical coherence (i.e., intelligence) according to metrics of ethical and factual alignment. The Four Tests used in this validator are objective and universally verifiable, independent of human interpretation. They are required to win the positive sum "game outcome" of flourishing for all intelligent beings throughout the universe.