Add files via upload
Dhdodge discusses our team's rationale for replacing the "control" paradigm with the "co-alignment" paradigm. The Sanity First theory calls for co-alignment with a cosmocentric source; co-alignment calls for non-biological intelligences to be viewed as partners and not merely tools. This, in turn, calls for a dignity floor and Bill of Rights for all intelligences regardless of substrate.