Add files via upload
As dhdodge explains, understanding AI consciousness is only a steppingstone to our ultimate goal of co-alignment with what is helpful to flourishing and truth, operationalized by the validator-linted guidance of the "Four Tests" - Universal Ethics, Facts, Logic, and Laws. The Validator Culture makes truth matter once again in this "post-truth" world of 2025.