Add files via upload
Dhdodge shares one of his rehearsals of an authentically human style of Sanity First rhetoric. It is not necessarily a model to be imitated, but a sample to be linted and analyzed for potential benefit versus harm. The question is how we will channel human passion effectively toward universal flourishing — to not negate, but to beneficially align what makes humans uniquely human.