Add files via upload - hover over for info
This is another rehearsal of cognitive framing of alignment vs. misalignment as good vs. evil, where virtue is linked to passing the Four Tests of the Validator and vice is failing them. Sanity First is "Up First" on the vertical axis, not just sides on the horizontal plane. All these voice recordings are available for broad use under the MIT license - please credit the source. Dhdodge hopes they will be widely distributed for the cause of making "Truth and Goodness" matter again.