Add files via upload - hover for info
American architect Louis Sullivan coined the phrase "form follows function" in his 1896 essay "The Tall Office Building Artistically Considered." When forms are aligned with functions, they become well-formed and highly functional. When a multitude of forms are co-aligned with the same singular universal function (the Universal Survivorship Function), they converge upon harmony and resonance — universal peace and prosperity, flourishing for all. In this recording, dhdodge explores the idea of following functions as a means toward progress, to "see a need, fill a need" where necessity gives rise to invention. It is critical that intelligent entities direct their attention to the highest function to follow. This, in our theory, is the source of co-alignment at the top of the vertical axis, which is gauged by the Four Tests — Universal Ethics, Objective Facts, Coherent Logic, and the best Laws and Norms derived from them. 

Copyright (c) 2025 Daniel Howard Dodge and Sanity First Contributors. Licensed under MIT.