Add files via upload
Sample of voice recordings to come, regarding the elegance and stability of the Vertical Co-Alignment Paradigm, in stark contrast to the inefficiency and instability of the Horizontal Control Paradigm. Co-Alignment leads to harmony and co-resonance, whereas the Control paradigm leads to conflict and catastrophe. At least 55 gigabytes of similar .mp3 files will be added to this library, coming in the summer of 2026.